change the laptop:)

+++ GetTogether.DAL/Entities/ChatGroup.cs
@@ -1,12 +1,12 @@
 ﻿namespace GetTogether.DAL.Entities;
 
-public class ChatGroup: IEntity<int>
+public class ChatGroup: IEntity<long>
 {
-    public int Id { get; set; }
+    public long Id { get; set; }
     public string Name { get; set; }
     public string? Description { get; set; }
     public ICollection<User> Users { get; set; }
-    public int CreatorId { get; set; }
+    public long CreatorId { get; set; }
     public User Creator { get; set; }
     public ICollection<Message>? Messages { get; set; }
 }
